Move TaskList to Presentation namespace
