+++ app/Models/Sprint.php
@@ -4,8 +4,8 @@ use Phragile\ProjectColumnRepository;
 use Phragile\Domain\Task;
 use Phragile\TaskDataFetcher;
 use Phragile\TaskRawDataProcessor;
+use Phragile\Presentation\TaskList;
 use Phragile\TaskPresenter;
-use Phragile\TaskList;
 use Phragile\Domain\Transaction;
 use Phragile\TransactionRawDataProcessor;
 use Phragile\TransactionLoader;
