+++ app/Phragile/BurndownChart.php
@@ -1,6 +1,7 @@
 <?php
 namespace Phragile;
 
+use Phragile\Presentation\TaskList;
 use Phragile\Domain\Transaction;
 
 class BurndownChart {
